# genai/api/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: contact@example.com
    url: http://example.com/contact
  description: "This is an **example** API to demonstrate features of OpenAPI specification\n\
    # Introduction\nThis API definition is intended to to be a good starting point\
    \ for describing your API in \n[OpenAPI/Swagger format](https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md).\n\
    It also demonstrates features of [create-openapi-repo](https://github.com/Redocly/create-openapi-repo)\
    \ tool and \n[Redoc](https://github.com/Redocly/Redoc) documentation engine. So\
    \ beyond the standard OpenAPI syntax we use a few \n[vendor extensions](https://github.com/Redocly/Redoc/blob/master/docs/redoc-vendor-extensions.md).\n\
    \n# OpenAPI Specification\nThe goal of The OpenAPI Specification is to define\
    \ a standard, language-agnostic interface to REST APIs which\nallows both humans\
    \ and computers to discover and understand the capabilities of the service without\
    \ access to source\ncode, documentation, or through network traffic inspection.\
    \ When properly defined via OpenAPI, a consumer can \nunderstand and interact\
    \ with the remote service with a minimal amount of implementation logic. Similar\
    \ to what\ninterfaces have done for lower-level programming, OpenAPI removes the\
    \ guesswork in calling the service.\n"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://example.com/terms/
  title: Example.com
  version: 1.0.0
  x-logo:
    url: https://redocly.github.io/openapi-template/logo.png
externalDocs:
  description: Find out how to create a GitHub repo for your OpenAPI definition.
  url: https://github.com/Rebilly/generator-openapi-repo
servers:
- url: http://example.com/api/v1
- url: https://example.com/api/v1
tags:
- description: Example echo operations
  name: Echo
- description: Operations about user
  name: User
paths:
  /users/{username}:
    get:
      description: "Some description of the operation. \nYou can use `markdown` here.\n"
      operationId: getUserByName
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      - description: The name that needs to be fetched
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      - description: Filter users without email
        explode: true
        in: query
        name: with_email
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                username: user1
                email: user@example.com
              schema:
                $ref: '#/components/schemas/User'
          description: Success
        "403":
          description: Forbidden
        "404":
          description: User not found
      security:
      - main_auth:
        - read:users
      - api_key: []
      summary: Get user by user name
      tags:
      - User
    put:
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
      - description: Pretty print response
        explode: true
        in: query
        name: pretty_print
        required: false
        schema:
          type: boolean
        style: form
      - description: The name that needs to be updated
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
        description: Updated user object
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Invalid user supplied
        "404":
          description: User not found
      security:
      - main_auth:
        - write:users
      summary: Updated user
      tags:
      - User
  /echo:
    post:
      description: Receive the exact message you've sent
      operationId: echo
      requestBody:
        content:
          application/json:
            schema:
              example: Hello world!
              type: string
          application/xml:
            schema:
              example: Hello world!
              type: string
        description: Echo payload
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                response:
                  value: Hello world!
              schema:
                type: string
            application/xml:
              schema:
                type: string
            text/csv:
              schema:
                type: string
          description: OK
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              explode: false
              schema:
                format: int32
                type: integer
              style: simple
            X-Expires-After:
              $ref: '#/components/headers/ExpiresAfter'
      security:
      - api_key: []
      - basic_auth: []
      summary: Echo test
      tags:
      - Echo
components:
  headers:
    ExpiresAfter:
      description: date in UTC when token expires
      explode: false
      schema:
        format: date-time
        type: string
      style: simple
  schemas:
    Email:
      description: User email address
      example: john.smith@example.com
      format: test
      title: email
      type: string
    User:
      example:
        firstName: John
        lastName: Smith
        email: john.smith@example.com
        username: John78
      properties:
        username:
          description: User supplied username
          example: John78
          minLength: 4
          title: username
          type: string
        firstName:
          description: User first name
          example: John
          minLength: 1
          title: firstName
          type: string
        lastName:
          description: User last name
          example: Smith
          minLength: 1
          title: lastName
          type: string
        email:
          description: User email address
          example: john.smith@example.com
          format: test
          title: email
          type: string
      title: User
      type: object
  securitySchemes:
    main_auth:
      flows:
        implicit:
          authorizationUrl: http://example.com/api/oauth/dialog
          scopes:
            read:users: read users info
            write:users: modify or remove users
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey
    basic_auth:
      scheme: basic
      type: http